requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.16.58
botocore==1.19.58
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.1.5
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
jmespath==0.10.0
lxml==4.6.2
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
style==1.1.0
urllib3==1.26.2
